print size of vars

+++ 0x00-hello_world/6-size.c
@@ -7,8 +7,8 @@ int main(void)
 {
 	printf("Size of a char: %lu byte(s) \n", (unsigned long)sizeof(char));
 	printf("Size of an int: %d byte(s)\n", sizeof(int));
-	printf("Size of a long int: %lu byte(s) \n",(unsigned long)sizeof(int));
-	printf("Size of a long long-int: %lu byte(s) \n",(unsigned long long)sizeof(int));
-	printf("Size of a float: %lu byte(s) \n",(unsigned long) sizeof(float));
+	printf("Size of a long int: %lu byte(s) \n", (unsigned long)sizeof(int));
+	printf("Size of a long long-int: %llu byte(s) \n", (unsigned long)sizeof(int));
+	printf("Size of a float: %lu byte(s) \n", (unsigned long) sizeof(float));
 	return (0);
 }
